requirements.txt
django==4.2
jp2forge==0.9.1
celery==5.3.1
redis==4.6.0
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
django-celery-results==2.5.1
Pillow==10.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
